refactor: add new 'con sole' pkg with principalidentity helpers
